requirements.txt
openai
langchain
faiss-cpu
qdrant-client
neo4j
py2neo
rdflib
transformers
sentence-transformers
loguru
